Clean up the read/write of shockwave refinement parameters, change level from float to int.
--HG--
branch : week-of-code

+++ src/enzo/global_data.h
@@ -521,7 +521,7 @@ EXTERN float RefineByResistiveLengthSafetyFactor;
 
 EXTERN float ShockwaveRefinementMinMach;
 EXTERN float ShockwaveRefinementMinVelocity;
-EXTERN float ShockwaveRefinementMaxLevel;
+EXTERN int ShockwaveRefinementMaxLevel;
 
 /* Noh problem switch: Upper-Right quadrant or full domain */
